refactor(query-builder): update code to reference v2 instead of v1
