+++ frontend/src/hooks/queryBuilder/useGetAggregateKeys.ts
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useQuery, UseQueryOptions, UseQueryResult } from 'react-query';
 import { getAggregateKeys } from 'api/queryBuilder/getAttributeKeys';
 import { QueryBuilderKeys } from 'constants/queryBuilder';
-import { InfraMonitoringEntity } from 'container/InfraMonitoringK8s/constants';
+import type { InfraMonitoringEntity } from 'container/InfraMonitoringK8sV2/constants';
 import { ErrorResponse, SuccessResponse } from 'types/api';
 import { IGetAttributeKeysPayload } from 'types/api/queryBuilder/getAttributeKeys';
 import { IQueryAutocompleteResponse } from 'types/api/queryBuilder/queryAutocompleteResponse';
